requirements.txt
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.3.0
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
bcrypt==3.2.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.1
colorama==0.4.4
coverage==5.1
cryptography==3.2.1
dictdiffer==0.8.1
dynamics365crm-python==0.1.2
entrypoints==0.3
Flask==1.1.2
greenlet==1.1.0
idna==2.9
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
json2xml==3.0.1
jsonschema==3.2.0
kaleido==0.0.3.post1
lxml==4.5.0
Mako==1.1.2
Markdown==2.4.1
MarkupSafe==1.1.1
msal==1.12.0
mss==5.0.0
multi-key-dict==2.0.3
numpy==1.18.3
packaging==20.9
pandas==1.0.3
paramiko==2.7.2
pbr==5.4.5
pdfrw==0.4
pdoc==0.3.2
Pillow==7.1.2
ping3==2.6.6
playwright==1.12.1
plotly==4.11.0
pluggy==0.13.1
portpicker==1.3.1
pushbullet.py==0.11.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8.1
pyee==8.1.0
PyJWT==2.1.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyodbc==4.0.30
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-http-client==3.2.7
python-jenkins==1.7.0
python-magic==0.4.18
pytz==2019.3
reportlab==3.5.53
requests==2.23.0
requests-toolbelt==0.9.1
retrying==1.3.3
scp==0.13.3
selenium==3.141.0
sendgrid==6.3.1
sentry-sdk==0.19.5
six==1.14.0
smartsheet-python-sdk==2.105.1
soupsieve==2.0.1
toml==0.10.2
toolz==0.11.1
tornado==6.0.4
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.25.9
websocket-client==0.57.0
websockets==9.1
Werkzeug==1.0.1
xlrd==1.2.0
xmldiff==2.4
xmlrunner==1.7.7
gnupg==1.3
